requirements.txt
openai
python-dotenv
pypdf
psycopg[binary] # PostgreSQL接続用ドライバ
pgvector # pgvectorのPythonバインディング